Make Feed UI element

+++ src/types/feeds.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct FeedMeta {
-    pub index: usize,
+    pub index: i32,
     pub feed_url: String,
     pub xml_file_path: Option<String>,
 }
